[DSPDC-1791] Expose helm chart linting as a task (#22)
Why

During testing of the YAML modifications for DSPDC-1791, I bumped into invalid YAML several times. We have a lint task for this, but it requires a full test suite run as it's attached at the test scope. It would be useful to be able to lint the chart without a full test suite run.

This PR

Creates a new lintHelmChart task key and assigns the helm chart linting code to it so it's invocable by external projects like HCA
"Aliases" the test task to the lintHelmChart so it runs by default during test runs as well
Upgrades SCIO + Beam